requirements.txt
appdirs
argparse
python-mpd2
musicbrainzngs
requests
urllib3
